[NFC] Share argument order and names between the arguments and results code
This was an annoying change, but mostly mechanical.  The goal here is
to create parallel structure between the two paths so that it isn't
utterly confusing to share common CRTP paths between them.  The renaming
of translate->process is just to make sure that all of the call sites
get correctly rewritten.