Fix argument side effects when an argument’s only user is a partial_apply which is returned
In the current SideEffectAnalysis, if an argument's only user is a partial_apply which
in turn is returned from the function, memory effect none is returned.

Since the returned partial_apply can be later applied via the return value, we have to be conservative here.

This resulted in a store being deleted incorrectly in DeadStoreElimination.

Fixes rdar://113339972