Creates new autodiff loop context allocation builtins and preserves old ones
This commit builds on top of a previous commit that fixed memory leak
issues in the autodiff loop context allocation builtins. The previous
commit changed the existing builtins. However, inorder to maintain ABI
compatibility and not break other consumers of autodiff, this commit
creates newly named builtins that don't leak memory and preserves the
older builtins.

As a result, now, code compiled with an older frontend and using a newer
stdlib will continue to work. But code compiled with a newer frontend
and using older stdlib will have undefined behavior.