[consume-operator] Use SSAPrunedLiveness rather than DiagnosticPrunedLiveness.
I also deleted DiagnosticPrunedLiveness. As part of this, I added a
post-processing step that finds non consuming uses that are in LiveOut regions
so that we can search for them when searching for reachable uses. This can
happen if the non consuming use is in a loop (implying it isn't on the
boundary), but is actually the thing we want to error on.

As a nice benefit, we also began to emit a note that we were missing previously
in consume_operator_kills_copyable_values.swift.

rdar://113142446