Frontend: introduce a new `swift_upcoming` search path entry
This adds `/usr/lib/swift_upcoming` to the default search path that
follows `/usr/lib/swift` to allow us to have a staging ground for
upcoming features.  Given that upcoming features may be ABI unstable,
they are statically linked.  The default behaviour of isolating static
and dynamic libraries into `/usr/lib/swift` and `/usr/lib/swift_static`
and switching that based upon `-static-stdlib` is important to preserve
as the linkage is encoded into the swift module and impacts the code
generation on some platforms.  This allows us to include a default
search path that may link statically irrespective of how the remainder
of the standard library is linked.