[AutoDiff] Handle init_enum_data_addr and inject_enum_addr for Optional
Optional's `init_enum_data_addr` and `inject_enum_addr` instructions are
generated in presence of non-loadable Optional values. The compiler used to
treat these instructions as inactive, and this resulted in silent run-time
issues described in #64223.

The patch marks `init_enum_data_addr` as "active" if its Optional operand is
also active, and in PullbackCloner we differentiate through it and the related
`inject_enum_addr`.

However, we only determine this relation in simple cases when both instructions
are in the same block. There is no def-use relation between them (both take the
same Optional operand), so if there is more than one set of instructions
operating on the same Optional, or there is some control flow, we currently bail
out.

In PullbackCloner, we walk over instructions in reverse order and start from
`inject_enum_addr` and its `Optional<Wrapped>.TangentVector` operand. Assuming
that is is already initialized, we emit an `unchecked_take_enum_data_addr` and
set it as the adjoint buffer of `init_enum_data_addr`. The Optional value is
invalidated, and we have to destroy the enum data address later when we reach
`init_enum_data_addr`.

Before this patch, `Optional<Wrapped>.TangentVector` was implemented as a struct
with an `Optional<Wrapped.TangentVector>` field (`value`). In order to use
`unchecked_take_enum_data_addr`, we have to access the underlying Optional
field, but it is not possible, because the struct is resilient:

    SIL verification failed: cannot access storage of resilient struct:
      !checkResilience(sd, F.getModule().getSwiftModule(), F.getResilienceExpansion())
    Verifying instruction:
	 %6 = alloc_stack $Optional<τ_0_0>.TangentVector
    ->   %8 = struct_element_addr %6 : $*Optional<τ_0_0>.TangentVector, #Optional.TangentVector.value
	 %9 = unchecked_take_enum_data_addr %8 : $*Optional<τ_0_0.TangentVector>, #Optional.some!enumelt

Therefore, the patch changes `Optional<Wrapped>.TangentVector` from struct to
enum with same fields as Optional (`.some` and `.none`). This way we use
`unchecked_take_enum_data_addr` on the `TangentVector` directly.