[cxx-interop][SwiftToCxx] Emit foreign reference types with `*` in the generated Clang header
C++ foreign reference types are imported to Swift as classes, not structs. When generating a Clang header that represents a Swift struct holding a C++ FRT as a field, we need to print the FRT as a pointer.

This was previously hitting an assertion in debug builds of the compiler: `assert(CD->isObjC())`.

rdar://114711899
(cherry picked from commit c45753dca2253e924a86d00b6117ee21ca41722c)