Basic: avoid truncation
`Swift.Int` is *NOT* `int` or `long`, it is `uintptr_t`.  Unfortunately,
we cannot include `stdint.h` here and thus cannot use `uintptr_t`.
Perhaps we should consider using `__swift_uintptr_t`.

For now, assume that we are always 64-bit and use `long long` and
`unsigned long long` which repairs the builds with non-LP64 targets.