tests: replace shell with python for portability
Replace the process_fine_grained_swiftdeps.sh with a python equivalent
(which also preserves the horrendous handling of YAML and even
"faithfully" replicates the horrible global variables).  This enables a
number of tests on Windows although the instigating macro test is not
yet enabled due to the need for further tweaks to the tests.