ASTGen: add a workaround for CSC conformance visibility
A thrown error is stored as `any Error` which does not conform to
`CustomStringConvertible`.  When we perform the conversion via the
`String(describing:)` initialiser, for some reason the runtime does not
find the conformance and the error is not translated properly.
Explicitly casting to `PluginError` resolves the missing conformance and
fixes a test failure.